+++ src/components/Portal/Portal.tsx
@@ -3,7 +3,7 @@
 import * as React from "react";
 import { createPortal } from "react-dom";
 
-export default function Modal({ children }: React.PropsWithChildren) {
+export default function Portal({ children }: React.PropsWithChildren) {
   const [mounted, setMounted] = React.useState(false);
 
   React.useEffect(() => setMounted(true), []);
