+++ src/data/footer.tsx
@@ -115,38 +115,38 @@ export const footer = {
     {
       id: "17",
       title: "Linkedin",
-      href: "/",
-      icon: <LinkedinIcon />,
+      href: "#",
+      icon: LinkedinIcon,
     },
     {
       id: "18",
       title: "Facebook",
-      href: "/",
-      icon: <FacebookIcon />,
+      href: "#",
+      icon: FacebookIcon,
     },
     {
       id: "19",
       title: "Instagram",
-      href: "/",
-      icon: <InstagramIcon />,
+      href: "#",
+      icon: InstagramIcon,
     },
     {
       id: "20",
       title: "Skype",
-      href: "/",
-      icon: <SkypeIcon />,
+      href: "#",
+      icon: SkypeIcon,
     },
     {
       id: "21",
       title: "Twitter",
-      href: "/",
-      icon: <TwitterIcon />,
+      href: "#",
+      icon: TwitterIcon,
     },
     {
       id: "22",
       title: "Pinterest",
-      href: "/",
-      icon: <PinterestIcon />,
+      href: "#",
+      icon: PinterestIcon,
     },
   ],
 };
